refactor(mapgen): finalize foundation artifact ownership (#2025)
refactor(mapgen): finalize foundation artifact ownership

Move Foundation artifact ownership into direct domain artifact files, route recipe publication through the public Foundation domain surface, narrow stage validation to tile/projection artifacts, and convert Resources direct artifacts to the artifacts/*.artifact.ts shape. Update Habitat preservation to protect domain artifact authority and stage wiring.

Verification: nx run mod-swooper-maps:check; nx run mod-swooper-maps:test; nx run mod-swooper-maps:habitat:check; targeted foundation artifact/guard tests; final agent review loop.

refactor(mapgen): normalize remaining artifact surfaces

refactor(mapgen): consolidate artifact validation owners

docs(habitat): clarify domain config prework law

docs(habitat): add domain config law investigation corpus

docs(habitat): draft domain config law execution sequence

docs(habitat): add config law hardening pattern gates

feat(habitat): add config law destination patterns

docs(habitat): add config law second-pass brief

docs(habitat): finalize config law disposition corpus

docs(habitat): mark config law disposition authority

docs(habitat): expand config law execution draft scope

docs(habitat): consolidate config law disposition source

docs(habitat): harden config law execution workstream

docs(habitat): enforce domain config repair rails

Add the active repair packet for the Domain Model Config Law sweep, including the red ledger, repair execution plan, reviewed agent prompts, and enforcement findings.\n\nRename the operation and recipe-stage file-shape Habitat rules into normalized rule IDs, add the advisory domain source topology rule, and supersede the historical execution packet so the current repair state is unambiguous.\n\nVerification:\n- bun habitat classify .habitat/.active\n- bun habitat classify .habitat/blueprints\n- bun habitat classify .habitat/scopes\n- bun habitat check --rule require_recipe_stage_authoring_file_shape --json --output /tmp/habitat-red-experiment/stage-authoring.json || true\n- bun habitat check --rule require_domain_operation_contract_file_shape --json --output /tmp/habitat-red-experiment/op-contract.json || true\n- bun habitat check --rule require_domain_source_topology --json --output /tmp/habitat-red-experiment/domain-topology.json || true\n- bun habitat check --rule preserve_standard_stage_topology_and_path_invariants --json --output /tmp/habitat-red-experiment/standard-stage-topology.json || true\n- git diff --check -- .habitat

+++ mods/mod-swooper-maps/test/ecology/feature-habitat-eligibility.test.ts
@@ -17,7 +17,7 @@ describe("ecology feature habitat eligibility", () => {
     const openOceanMask = new Uint8Array([0, 1, 0, 0]);
     const lakeMask = new Uint8Array([1, 0, 0, 0]);
     const coastalWater = new Uint8Array([1, 0, 1, 1]);
-    const distanceToCoast = new Uint16Array([1, 5, 1, 1]);
+    const distanceToCoast = new Uint8Array([1, 5, 1, 1]);
 
     const reef = ecology.ops.scoreReef.run(
       {
